add bathroom model unit tests [#172282259]

+++ aspnet/RVTR.Lodging.ObjectModel/Models/BathroomModel.cs
@@ -9,7 +9,7 @@ namespace RVTR.Lodging.ObjectModel.Models
   public class BathroomModel : IValidatableObject
   {
     public int Id { get; set; }
-    public int Fixture { get; set; }
+    public double Fixture { get; set; }
 
     /// <summary>
     /// Represents the _Bathroom_ `Validate` method
